Update README.md
# Data_Analysis_Portfolio_Project
Project conducted using Jupyter Notebook and shared as part of the porfolio. Traffic patterns on Hwy 6 in Calvert, Texas, where our family bookstore is was chosen. 
Data was gleaned from the TPP District Traffic Web Viewer. The 2021 AADT data set is prepared by the Texas Department of Transportation, Transportation Planning and Programming Division, Transportation System Support Branch in cooperation with the United States Department of Transportation Federal Highway Administration.
 
Copyright 2021 by the Texas Department of Transportation. All rights reserved.
 
Notice: This data was produced for internal use and is provided for informational purposes only, and subject to change.  TxDOT makes no warranty of any kind as to the accuracy or validity of the data provided. GIS content and data provided is the property of TxDOT and may not be used for commercial purposes or resold. Distribution of this content to third parties without TxDOT's written consent is strictly prohibited.
 
Questions and Comments: 512-486-5000
or TPP-Traffic-Data@txdot.gov

Publication Data: July 2022